+++ Runtime/Server/DedicatedServerManager.cs
@@ -41,6 +41,15 @@ namespace AccelByte.Server
         }
 
         /// <summary>
+        /// Get the registered server name.
+        /// will be filled after server is registered.
+        /// </summary>
+        public string ServerName
+        {
+            get { return serverName; }
+        }
+
+        /// <summary>
         /// Register server to DSM and mark this machine as ready
         /// </summary>
         /// <param name="port">Exposed port number to connect to</param>
@@ -143,6 +152,8 @@ namespace AccelByte.Server
             , ResultCallback callback )
         {
             Report.GetFunctionLog(GetType().Name);
+
+            serverName = inName;
 
             coroutineRunner.Run(api.RegisterLocalServer(port, inName, callback));
         }